working on correcting \NimbusApp\Nimbus2\Nimbus\Nimbus\Nimbus.Web.Client\obj\Debug\net8.0\Nimbus.Web.Client.MvcApplicationPartsAssemblyInfo.cs(14,54,14,78): error CS0234: The type or namespace name 'ApplicationPartAttributeAttribute' does not exist in the namespace 'Microsoft.AspNetCore.Mvc.ApplicationParts'

+++ Nimbus/Nimbus.Shared/PagesCodeBehind/ViewRoute.cs
@@ -1,5 +1,7 @@
-﻿using Nimbus.Shared.Entities;
+﻿using Microsoft.AspNetCore.Components;
+using Nimbus.Shared.Entities;
 using Nimbus.Shared.Repositories;
+using Nimbus.Shared.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -7,41 +9,59 @@ using System.Reflection.Metadata.Ecma335;
 using System.Text;
 using System.Threading.Tasks;
 
+
 namespace Nimbus.Shared.Pages
 {
    
-    partial class ViewRoute
+partial class ViewRoute : ComponentBase
     {
+        //[Inject]
+        //public SelectionService SelectionService { get; set; }
+
+        //[Inject]
+        //public IRouteRepository RouteRepository { get; set; }
+
+        //[Inject]
+        //public IAddressRepository AddressRepository { get; set; }
+
         public int routeId;
         public bool editAddress = false;
         public bool deleteAddress = false;
         Address? address;
-        //RouteEntity route;
         List<Address> allStops = new List<Address>();
+
         protected override async Task OnInitializedAsync()
         {
             if (SelectionService.selectedRoute != null)
             {
-                allStops = await Task.Run(() => RouteRepository.GetStopsAsync(SelectionService.selectedRoute.Id));
+                allStops = await RouteRepository.GetStopsAsync(SelectionService.selectedRoute.Id);
             }
         }
+
         public async Task DeleteAddressFinal()
         {
-            await AddressRepository.DeleteAddressAsync(address.id);
-            deleteAddress = false;
+            if (address != null)
+            {
+                await AddressRepository.DeleteAddressAsync(address.id);
+                deleteAddress = false;
+            }
         }
-        
+
         public async Task EditAddress(int id)
         {
             editAddress = true;
             address = await AddressRepository.FindAddressForRouteByIdAsync(SelectionService.selectedRoute.Id, id);
-
         }
+
         public async Task SaveAddressAsync()
         {
-            await AddressRepository.UpdateAddressAsync(address);
-            editAddress = false;
+            if (address != null) // Check for null reference
+            {
+                await AddressRepository.UpdateAddressAsync(address);
+                editAddress = false;
+            }
         }
+
         public async Task DeleteAddress(int id)
         {
             deleteAddress = true;
@@ -53,6 +73,7 @@ namespace Nimbus.Shared.Pages
         {
             editAddress = false;
         }
+
         private void CloseDeleteModal()
         {
             deleteAddress = false;
